backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
sqlalchemy
weasyprint
pydantic
pyyaml
langdetect